web_gateway: carve input_authority.rs (pure move)
The display input-authority domain moves to
web_gateway/input_authority.rs: the DisplayInputHolder state machine,
grant/release application for local, dashboard-control, and federated
websockets, federated authority subscribers, and the bootstrap
snapshots — with its 49 tests and 4 test helpers. seed_holder stays as
a shared fixture; DisplayInputAuthorityChange's fields widen to
pub(crate) for the listener.
